clean up PocketAppDelegate
